Refactored pause menu and added leave button
